backend/config/database.py
import asyncpg
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = f"postgresql://{os.getenv('DB_USER', 'postgres')}:{os.getenv('DB_PASSWORD')}@{os.getenv('DB_HOST', 'localhost')}:{os.getenv('DB_PORT', '5432')}/{os.getenv('DB_NAME', 'academia_final')}"

pool = None

async def get_db_pool():
    global pool
    if pool is None:
        pool = await asyncpg.create_pool(
            DATABASE_URL,
            min_size=5,
            max_size=20,
            command_timeout=60
        )
    return pool

async def close_db_pool():
    global pool
    if pool:
        await pool.close()
        pool = None

async def get_db():
    pool = await get_db_pool()
    async with pool.acquire() as connection:
        yield connection
